SignTextComposite: Further unhorrification, fix relations on sentences
Before this, "relating objects", i.e. things that have relations to
something that is part of a text, would only be handled correctly if the
target was a word or marker. Now cases targeting an Ambivalence,
a LemmaCase inside one and whole BTSSentences also work.

Also, selections on markers now highlight the proper comments.